Move general video context tests to the new folder structure
