+++ src/Context/Video/Infrastructure/Symfony/Bundle/DependencyInjection/Resources/infrastructure/video_arranger.yml
@@ -1,6 +1,6 @@
 services:
 
   codely.video.infrastructure.arranger:
-    class: CodelyTv\Context\Video\Test\Arranger\VideoEnvironmentArranger
+    class: CodelyTv\Test\Context\Video\Arranger\VideoEnvironmentArranger
     arguments:
       - '@codely.video.infrastructure.database'
