+++ src/Context/Video/Module/User/Test/PhpUnit/UserModuleFunctionalTestCase.php
@@ -2,7 +2,7 @@
 
 namespace CodelyTv\Context\Video\Module\User\Test\PhpUnit;
 
-use CodelyTv\Context\Video\Test\PhpUnit\VideoContextFunctionalTestCase;
+use CodelyTv\Test\Context\Video\PhpUnit\VideoContextFunctionalTestCase;
 
 abstract class UserModuleFunctionalTestCase extends VideoContextFunctionalTestCase
 {
