--- src/Context/Video/Test/PhpUnit/VideoContextUnitTestCase.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace CodelyTv\Context\Video\Test\PhpUnit;
-
-use CodelyTv\Test\PhpUnit\TestCase\Module\ModuleUnitTestCase;
-
-abstract class VideoContextUnitTestCase extends ModuleUnitTestCase
-{
-}
