+++ tests/Context/Video/Module/Notification/NotificationModuleFunctionalTestCase.php
@@ -3,7 +3,7 @@
 namespace CodelyTv\Test\Context\Video\Module\Notification;
 
 use CodelyTv\Context\Video\Module\Notification\Domain\NotificationRepository;
-use CodelyTv\Context\Video\Test\PhpUnit\VideoContextFunctionalTestCase;
+use CodelyTv\Test\Context\Video\VideoContextFunctionalTestCase;
 
 abstract class NotificationModuleFunctionalTestCase extends VideoContextFunctionalTestCase
 {
