+++ tests/Context/Video/VideoContextUnitTestCase.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace CodelyTv\Test\Context\Video;
+
+use CodelyTv\Test\Infrastructure\PHPUnit\Module\ModuleUnitTestCase;
+
+abstract class VideoContextUnitTestCase extends ModuleUnitTestCase
+{
+}
