+++ tests/Infrastructure/Bus/Command/CommandBusSyncTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace CodelyTv\Tests\Infrastructure\Bus\Command;
+namespace CodelyTv\Test\Infrastructure\Bus\Command;
 
 use CodelyTv\Infrastructure\Bus\Command\CommandBusSync;
 use CodelyTv\Shared\Domain\Bus\Command\Command;
-use CodelyTv\Test\PhpUnit\TestCase\UnitTestCase;
+use CodelyTv\Test\Infrastructure\PHPUnit\UnitTestCase;
 use Mockery\MockInterface;
 use RuntimeException;
